tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/133.html

<!DOCTYPE html>
<html><head>
	<title>scheduler: inline HTML script added by SVG script </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test();
  log("inline script #1")
</script>
<svg>
<script>
log("inline script #2")
var s = document.createElement("script");
s.textContent = "log('inline script #3');";
document.getElementsByTagName("svg")[0].appendChild(s);
log("end inline script #2");
</script>
</svg>
<script>
log("inline script #4");
t.step(function() {
  assert_array_equals(eventOrder, ["inline script #1", "inline script #2", "inline script #3", 
                                   "end inline script #2", "inline script #4"]);
  t.done();
});
</script>
